% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aggregate_indicators.R
\name{get_aggregate_indicators}
\alias{get_aggregate_indicators}
\title{retrieve aggregate indicator data}
\usage{
get_aggregate_indicators(symbol, resolution, api.key, write.file = FALSE)
}
\arguments{
\item{symbol}{the stock symbol to retrieve data for}

\item{resolution}{intervals for the data}

\item{api.key}{your API token from finnhub.io}

\item{write.file}{should the table be written to the "aggregate_indicators" folder?}
}
\value{
a data frame of aggregate indicators and trends
}
\description{
`get_aggregate_indicators` retrieves aggregate signal of multiple technical indicators (e.g. MACD, RSI, MA)
}
\examples{
\donttest{
### Get support and resistance levels with resolution of 1 minute
get_aggregate_indicators(symbol = 'AAPL', resolution = 1,
api.key = api.key)
}
}
