% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basic_financials.R
\name{get_basic_financials}
\alias{get_basic_financials}
\title{retrieve basic financial data}
\usage{
get_basic_financials(symbol, api.key, metric.type = "all", write.file = TRUE)
}
\arguments{
\item{symbol}{the stock symbol to retrieve data for}

\item{api.key}{your API token from finnhub.io}

\item{metric.type}{the type of metric to retrieve data for}

\item{write.file}{should the table be written to the "basic_financials" folder?}
}
\value{
a data frame of basic financial information for the specified company
}
\description{
`get_basic_financials` retrieves basic financial information for a given company
}
\examples{
\donttest{
### Get basic financial price data for Apple
### options: 'all', 'price', 'valuation', 'margin'
get_basic_financials(symbol = "AAPL", api.key = api.key, metric.type = "price")
}
}
