% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_filings.R
\name{get_company_filings}
\alias{get_company_filings}
\title{retrieve filings for a company}
\usage{
get_company_filings(
  symbol = "",
  cik = "",
  access.number = "",
  form = "",
  api.key,
  last.n.days = 180,
  start.date = NA,
  end.date = NA,
  write.file = FALSE
)
}
\arguments{
\item{symbol}{the stock symbol to retrieve company news for}

\item{cik}{the central index key to retrieve company filings for}

\item{access.number}{the access number to retrieve company filings for}

\item{form}{filter by type of filing form}

\item{api.key}{your API token from finnhub.io}

\item{last.n.days}{the range of days to get data for when start and end dates are NA (default)}

\item{start.date}{the earliest date to get news for (YYYY-MM-DD)}

\item{end.date}{the latest date to get news for (YYYY-MM-DD)}

\item{write.file}{should the table be written to the "filings" folder?}
}
\value{
a data frame of information on a company's filings
}
\description{
`get_company_filings` retrieves filings for a specified company or identifier
}
\examples{
\donttest{
### Get filings for Apple between the start of January 2020 and end of March 2020
get_company_filings(symbol = "AAPL", api.key = api.key,
start.date = '2020-01-01', end.date = '2020-03-31')
}
}
