% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_news.R
\name{get_company_news}
\alias{get_company_news}
\title{retrieve news about a company}
\usage{
get_company_news(
  symbol,
  api.key,
  start.date = NA,
  end.date = NA,
  last.n.days = 180,
  translate.date = TRUE,
  write.file = TRUE
)
}
\arguments{
\item{symbol}{the stock symbol to retrieve company news for}

\item{api.key}{your API token from finnhub.io}

\item{start.date}{the earliest date to get news for (YYYY-MM-DD)}

\item{end.date}{the latest date to get news for (YYYY-MM-DD)}

\item{last.n.days}{the range of days to get data for when startand end dates are NA (default)}

\item{translate.date}{should UNIX time be converted to standard date format?}

\item{write.file}{should a table of dates, headlines, sources, and urls be written to the "news" folder?}
}
\value{
a data frame of dates, headlines, sources, and urls
}
\description{
`get_company_news` retrieves news about a company for a given time range
}
\examples{
\donttest{
### Get news for Apple from the last 30 days
get_company_news(symbol = "AAPL", api.key = api.key, last.n.days = 30)

### Get news for Apple from March
get_company_news(symbol = "AAPL", api.key = api.key,
start.date = '2020-03-01', end.date = '2020-03-30')
}
}
