% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_peers.R
\name{get_company_peers}
\alias{get_company_peers}
\title{retrieve stock tickers supported by Finnhub}
\usage{
get_company_peers(symbol, api.key, write.file = TRUE)
}
\arguments{
\item{symbol}{the stock symbol to retrieve data for}

\item{api.key}{your API token from finnhub.io}

\item{write.file}{should the table be written to the "company_peers" folder?}
}
\value{
a data frame of symbols for related companies
}
\description{
`get_company_peers` retrieves a list of stocks with the same country and GICS sub-industry as the specified symbol
}
\examples{
\donttest{
### Get list of companies similar to Apple
get_company_peers(symbol = 'AAPL', api.key = api.key)
}
}
