% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_earnings_surprise.R
\name{get_earnings_surprise}
\alias{get_earnings_surprise}
\title{retrieve historical quarterly earnings surprise}
\usage{
get_earnings_surprise(symbol, api.key, num.periods = 4, write.file = TRUE)
}
\arguments{
\item{symbol}{the stock symbol to retrieve data for}

\item{api.key}{your API token from finnhub.io}

\item{num.periods}{the number of periods to get data for; leave blank to get full history}

\item{write.file}{should the table be written to the "earnings_surprise" folder?}
}
\value{
a data frame of actual earnings and estimates for the given company and time period
}
\description{
`get_earnings_surprise` retrieves company historical quarterly earnings surprise dating back to 2000
}
\examples{
\donttest{
### Get earnings surprises for the 4 most recent periods
get_earnings_surprise(symbol = 'AAPL', num.periods = 4, api.key = api.key)
}
}
