% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ipo_calendar.R
\name{get_ipo_calendar}
\alias{get_ipo_calendar}
\title{retrieve ipo calendar}
\usage{
get_ipo_calendar(
  api.key,
  last.n.days = 180,
  start.date = NA,
  end.date = NA,
  write.file = TRUE
)
}
\arguments{
\item{api.key}{your API token from finnhub.io}

\item{last.n.days}{the range of days to get data for when start and end dates are NA (default)}

\item{start.date}{the earliest date to get news for (YYYY-MM-DD)}

\item{end.date}{the latest date to get news for (YYYY-MM-DD)}

\item{write.file}{should the table be written to the "ipo_calendar" folder?}
}
\value{
a data frame of data on ipo dates
}
\description{
`get_ipo_calendar` retrieves information on ipo's for a given time period
}
\examples{
\donttest{
### Get ipo's from the last 30 days
get_ipo_calendar(api.key = api.key, last.n.days = 30)

### Get ipo's from March of 2020
get_ipo_calendar(api.key = api.key, start.date = '2020-03-01', end.date = '2020-03-31')
}
}
