% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_news.R
\name{get_market_news}
\alias{get_market_news}
\title{retrieve latest market news}
\usage{
get_market_news(
  category,
  api.key,
  minid = 0,
  translate.date = TRUE,
  write.file = FALSE
)
}
\arguments{
\item{category}{the category of news to retrieve}

\item{api.key}{your API token from finnhub.io}

\item{minid}{only get news after a specific id}

\item{translate.date}{should UNIX time be converted to standard date-time format?}

\item{write.file}{should the table be written to the "news" folder?}
}
\value{
a data frame of recent news articles for the user-specified category
}
\description{
`get_market_news` gets the latest market news for a given category
}
\examples{
\donttest{
### Get recent market news data
get_market_news(category = c("all", "general", "forex", "crypto", "merger"),
api.key = api.key)
}
}
