% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pattern_recognition.R
\name{get_pattern_recognition}
\alias{get_pattern_recognition}
\title{run pattern recognition on a symbol}
\usage{
get_pattern_recognition(symbol, resolution, api.key, write.file = TRUE)
}
\arguments{
\item{symbol}{the stock symbol to retrieve data for}

\item{resolution}{intervals for the data}

\item{api.key}{your API token from finnhub.io}

\item{write.file}{should the table be written to the "patterns" folder?}
}
\value{
a data frame of results from the pattern recognition algorithm
}
\description{
`get_pattern_recognition` runs a pattern recognition algorithm on a given stock
}
\examples{
\donttest{
### Run pattern recognition algorithm with resolution of 1 day for Apple
get_pattern_recognition(symbol = 'AAPL', resolution = 'D', api.key = api.key)
}
}
