% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quote.R
\name{get_quote}
\alias{get_quote}
\title{retrieve real-time stock data}
\usage{
get_quote(symbol, api.key, translate.date = TRUE, write.file = FALSE)
}
\arguments{
\item{symbol}{the stock symbol to retrieve data for}

\item{api.key}{your API token from finnhub.io}

\item{translate.date}{should UNIX time be converted to standard date-time format?}

\item{write.file}{should the table be written to the "quotes" folder?}
}
\value{
a data frame of prices for current, open, close, high, and low, along with UNIX timestamp
}
\description{
`get_quote` retrieves real-time data for a given stock
}
\examples{
\donttest{
### Get real-time quote for Apple
get_quote(symbol = 'AAPL', api.key = api.key)
}
}
