% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recommendation_trends.R
\name{get_recommendation_trends}
\alias{get_recommendation_trends}
\title{retrieve recommendation trends}
\usage{
get_recommendation_trends(symbol, api.key, write.file = FALSE)
}
\arguments{
\item{symbol}{the stock symbol to retrieve data for}

\item{api.key}{your API token from finnhub.io}

\item{write.file}{should the table be written to the "recommendation_trends" folder?}
}
\value{
a data frame of recommendation trends for the specified company
}
\description{
`get_recommendation_trends` retrieves recommendation trends for a given company
}
\examples{
\donttest{
### Get recommendation trends for Apple
get_recommendation_trends(symbol = "AAPL", api.key = api.key)
}
}
