% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reported_financials.R
\name{get_reported_financials}
\alias{get_reported_financials}
\title{retrieve reported financial data}
\usage{
get_reported_financials(
  symbol,
  api.key,
  frequency = "annual",
  write.file = TRUE
)
}
\arguments{
\item{symbol}{the stock symbol to retrieve data for}

\item{api.key}{your API token from finnhub.io}

\item{frequency}{the time period encompassed by the report(s)}

\item{write.file}{should the table be written to the "reported_financials" folder?}
}
\value{
a data frame of reported financial information for the specified company
}
\description{
`get_reported_financials` retrieves financial data through a company's filings
}
\examples{
\donttest{
### Get financial report data for Apple
get_reported_financials(symbol = "AAPL", api.key = api.key, frequency = c("annual", "quarterly"))
}
}
