% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentiments.R
\name{get_sentiments}
\alias{get_sentiments}
\title{retrieve sentiment analyses for a company}
\usage{
get_sentiments(symbol, api.key, write.file = TRUE)
}
\arguments{
\item{symbol}{the stock symbol to retrieve sentiment analyses for}

\item{api.key}{your API token from finnhub.io}

\item{write.file}{should the table be written to the "sentiments" folder?}
}
\value{
a data frame of sentiment analysis information
}
\description{
`get_sentiments` retrieves sentiment analyses of recent articles for a given company
}
\examples{
\donttest{
### Get recent sentiment analyses for Apple
get_sentiments(symbol = "AAPL", api.key = api.key)
}
}
