% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_candles.R
\name{get_stock_candles}
\alias{get_stock_candles}
\title{retrieve stock candles data}
\usage{
get_stock_candles(
  symbol,
  resolution,
  from.date,
  to.date,
  api.key,
  format = "csv",
  adjusted = FALSE,
  translate.input.dates = TRUE,
  translate.date = TRUE,
  write.file = FALSE
)
}
\arguments{
\item{symbol}{the stock symbol to retrieve data for}

\item{resolution}{intervals for the data}

\item{from.date}{the first date/time to retrieve data for}

\item{to.date}{the last date/time to retrieve data for}

\item{api.key}{your API token from finnhub.io}

\item{format}{the format of data to return (csv or json)}

\item{adjusted}{should adjusted data be returned?}

\item{translate.input.dates}{was the entered data in POSIXct format (YYYY-MM-DD hh:mm:ss timezone)?}

\item{translate.date}{should UNIX time be converted to standard date-time format?}

\item{write.file}{should the table be written to the "candles" folder?}
}
\value{
a data frame of stock candles for given inputs
}
\description{
`get_stock_candles` retrieves candlestick data for U.S. stocks dating as far back as 25 years
}
\examples{
\donttest{
### Get stock candles for Apple from last hour, in 1-minute intervals
to.date <- as.numeric(.POSIXct(Sys.time()))
from.date <- to.date - 3600

get_stock_candles(symbol = 'AAPL', api.key = api.key,
from.date = from.date, to.date = to.date)
}
}
