% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_splits.R
\name{get_stock_splits}
\alias{get_stock_splits}
\title{retrieve stock splits}
\usage{
get_stock_splits(symbol, from.date, to.date, api.key, write.file = FALSE)
}
\arguments{
\item{symbol}{the stock symbol to retrieve data for}

\item{from.date}{the first date to retrieve data for (YYYY-MM-DD)}

\item{to.date}{the last date to retrieve data for (YYYY-MM-DD)}

\item{api.key}{your API token from finnhub.io}

\item{write.file}{should the table be written to the "splits" folder?}
}
\value{
a data frame of stock splits
}
\description{
`get_stock_splits` retrieves stock splits for a given symbol
}
\examples{
\donttest{
### Get last 10 years of stock splits for Apple
from.date = '2010-01-01'
to.date = '2020-01-01'

get_stock_splits(symbol = 'AAPL', from.date = from.date,
to.date = to.date, api.key = api.key)
}
}
