% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_support_resistance.R
\name{get_support_resistance}
\alias{get_support_resistance}
\title{retrieve support and resistance data}
\usage{
get_support_resistance(symbol, resolution, api.key, write.file = FALSE)
}
\arguments{
\item{symbol}{the stock symbol to retrieve data for}

\item{resolution}{intervals for the data}

\item{api.key}{your API token from finnhub.io}

\item{write.file}{should the table be written to the "support_resistance" folder?}
}
\value{
an array of support and resistance levels
}
\description{
`get_support_resistance` retrieves support and resistance levels for a given company
}
\examples{
\donttest{
### Get support and resistance levels with resolution of 1 minute
get_support_resistance(symbol = 'AAPL', resolution = 1,
api.key = api.key)
}
}
