% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_supported_stocks.R
\name{get_supported_stocks}
\alias{get_supported_stocks}
\title{retrieve stock tickers supported by Finnhub}
\usage{
get_supported_stocks(exchange, api.key, write.file = TRUE)
}
\arguments{
\item{exchange}{the exchange to get list of symbols from}

\item{api.key}{your API token from finnhub.io}

\item{write.file}{should the table be written to the current working directory?}
}
\value{
a data frame of symbols, full company names, and descriptions
}
\description{
`get_supported_stocks` retrieves a list of stocks that are usable with Finnhub's API
}
\examples{
\donttest{
### Get list of supported stock symbols
get_stock_list(exchange = 'US', api.key = api.key)
}
}
