% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_price.R
\name{get_target_price}
\alias{get_target_price}
\title{retrieve latest price target consensus}
\usage{
get_target_price(symbol, api.key, write.file = TRUE)
}
\arguments{
\item{symbol}{the stock symbol to retrieve data for}

\item{api.key}{your API token from finnhub.io}

\item{write.file}{should the table be written to the "target_price" folder?}
}
\value{
a data frame of target prices based on the latest consensus
}
\description{
`get_target_price` retrieves the latest price target consensus for a given stock symbol
}
\examples{
\donttest{
### Get target prices for Apple
get_target_price(symbol = 'AAPL', api.key = api.key)
}
}
