% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_technical_indicators.R
\name{get_technical_indicators}
\alias{get_technical_indicators}
\title{retrieve technical indicator data}
\usage{
get_technical_indicators(
  symbol,
  resolution,
  from.date,
  to.date,
  api.key,
  indicator = "SMA",
  indicator.list = NA,
  translate.input.dates = TRUE,
  translate.date = TRUE,
  write.file = FALSE
)
}
\arguments{
\item{symbol}{the stock symbol to retrieve data for}

\item{resolution}{intervals for the data}

\item{from.date}{the first date/time to retrieve data for}

\item{to.date}{the last date/time to retrieve data for}

\item{api.key}{your API token from finnhub.io}

\item{indicator}{name of technical indicator (e.g. SMA, MACD, DX)}

\item{indicator.list}{list of indicator field parameters found on finnhub.io's documentation (arguments column)}

\item{translate.input.dates}{was the entered data in POSIXct format (YYYY-MM-DD hh:mm:ss timezone)?}

\item{translate.date}{should UNIX time be converted to standard date-time format?}

\item{write.file}{should the table be written to the "technical_indicators" folder?}
}
\value{
a data frame of technical indicators for the given company
}
\description{
`get_technical_indicators` retrieves technical indicator data for a given symbol
}
\examples{
\donttest{
### Get exponential moving average data for Apple over the past 5 hours with a resolution of 1-minute
to.date <- Sys.time()
from.date <- to.date - (5*3600)

get_technical_indicators(symbol = 'AAPL', resolution = 1,
from.date = from.date, to.date = to.date, api.key = api.key,
indicator = 'EMA')
}
}
