% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_finn_df.R
\name{write_finn_df}
\alias{write_finn_df}
\title{writes finnhub dataframe to csv}
\usage{
write_finn_df(dataframe, symbol, folder.name)
}
\arguments{
\item{dataframe}{a dataframe of data to write to file}

\item{symbol}{the stock symbol to write company news for}

\item{folder.name}{the name of the folder to write dataframe to}
}
\value{
success or failure to write file message
}
\description{
`write_finn_df` writes dataframe retrieved via finntools package to csv; mostly for internal use, as this is done by default
}
\examples{
\donttest{
### Use get_company_news to retrieve dataframe
AAPL.news <- get_company_news(symbol = 'AAPL', api.key = api.key)

### Write data to .csv file
write_finn_df(dataframe = AAPL.news, symbol = "AAPL", folder.name = 'news')
}
}
